Convert one container each time for `criu convert` command.
In previous implementation, I choose to let `criu convert` to work on
many containers. Because that method is easy to deal with page offset on
dax device.
But I think it is not flexible for a general tools like criu, so I
choose to let criu cares about only one container each time and add a
parameter called `--dax-pgoff`. After converting, it will output the
page num on dax device into a file called `convert-pgnum.img`.

The second thing is that I do not know why I cannot open /dev/pseudo_mm
while restoring. In qemu it always complains that "Operation not
  permitted". I come up with a workaround that let caller of criu to
  open it and pass the file descriptor to criu through inherit-fd.